src/utils/streaming.py
"""
Pathway streaming integration for real-time parking pricing system
"""

import pathway as pw
import pandas as pd
import numpy as np
from typing import Dict, Any, Optional, List
import json
import logging


class ParkingDataStream:
    """
    Real-time data streaming with Pathway for parking pricing system
    """
    
    def __init__(self, config: Optional[Dict] = None):
        """
        Initialize streaming pipeline
        
        Args:
            config: Streaming configuration
        """
        self.config = config or {
            'input_topic': 'parking_occupancy',
            'output_topic': 'pricing_decisions',
            'batch_size': 100,
            'window_size': '5min'
        }
        self.models = {}
        
    def setup_input_stream(self, input_source: str) -> pw.Table:
        """
        Setup input data stream from Pathway connector
        
        Args:
            input_source: Data source (Kafka, CSV, etc.)
            
        Returns:
            Pathway Table with streaming data
        """
        # Example: Kafka input stream
        if input_source.startswith('kafka://'):
            input_table = pw.io.kafka.read(
                rdkafka_settings={
                    'bootstrap.servers': input_source.replace('kafka://', ''),
                    'group.id': 'parking_pricing_consumer',
                    'auto.offset.reset': 'latest'
                },
                topic=self.config['input_topic'],
                format='json',
                schema=ParkingDataSchema
            )
        # Example: CSV file stream
        elif input_source.endswith('.csv'):
            input_table = pw.io.csv.read(
                input_source,
                schema=ParkingDataSchema,
                mode='streaming'
            )
        else:
            raise ValueError(f"Unsupported input source: {input_source}")
            
        return input_table
    
    def feature_engineering_pipeline(self, input_table: pw.Table) -> pw.Table:
        """
        Real-time feature engineering pipeline
        
        Args:
            input_table: Raw parking data stream
            
        Returns:
            Processed features table
        """
        # Add timestamp features
        features_table = input_table.with_columns(
            hour=input_table.timestamp.dt.hour(),
            day_of_week=input_table.timestamp.dt.dayofweek(),
            is_weekend=input_table.timestamp.dt.dayofweek() >= 5,
            is_rush_hour=((input_table.timestamp.dt.hour() >= 7) & 
                         (input_table.timestamp.dt.hour() <= 9)) |
                        ((input_table.timestamp.dt.hour() >= 17) & 
                         (input_table.timestamp.dt.hour() <= 19))
        )
        
        # Calculate occupancy rate
        features_table = features_table.with_columns(
            occupancy_rate=input_table.occupancy / input_table.capacity
        )
        
        # Add rolling statistics (e.g., last 1 hour average)
        features_table = features_table.with_columns(
            occupancy_avg_1h=input_table.occupancy.rolling_average(
                duration=pw.Duration.hours(1)
            )
        )
        
        return features_table
    
    def ml_inference_pipeline(self, features_table: pw.Table) -> pw.Table:
        """
        Real-time ML model inference pipeline
        
        Args:
            features_table: Processed features stream
            
        Returns:
            Predictions table
        """
        def predict_occupancy(row) -> float:
            """XGBoost occupancy prediction"""
            if 'xgboost_forecaster' in self.models:
                model = self.models['xgboost_forecaster']
                features = np.array([[
                    row['hour'], row['day_of_week'], row['is_weekend'], 
                    row['is_rush_hour'], row['occupancy_rate'], 
                    row['queue_length'], row['traffic_numeric'], 
                    row['vehicle_numeric'], row['is_special_day']
                ]])\n                try:\n                    return float(model.predict(features)[0])\n                except Exception as e:\n                    logging.error(f\"XGBoost prediction error: {e}\")\n                    return row['occupancy_rate']  # Fallback\n            return row['occupancy_rate']\n        \n        def predict_temporal_pattern(row) -> float:\n            \"\"\"LSTM/GRU temporal pattern prediction\"\"\"\n            if 'lstm_model' in self.models:\n                # Note: In practice, you'd need a sliding window of sequences\n                # This is a simplified version\n                return row['occupancy_rate']  # Simplified fallback\n            return row['occupancy_rate']\n        \n        def optimize_price(row) -> float:\n            \"\"\"Q-learning price optimization\"\"\"\n            if 'q_agent' in self.models:\n                # Get predicted occupancy from forecasting models\n                predicted_occupancy = row.get('predicted_occupancy', row['occupancy_rate'])\n                \n                # Create state vector for RL agent\n                state = np.array([\n                    predicted_occupancy, row['queue_length'], \n                    row['traffic_numeric'], row['is_rush_hour'],\n                    row['is_weekend'], row['hour'], row['day_of_week']\n                ])\n                \n                try:\n                    agent = self.models['q_agent']\n                    action = agent.predict(state.reshape(1, -1))[0]\n                    # Convert action to price (assuming price range 5-50)\n                    base_price = 10.0\n                    price_multiplier = 1.0 + (action * 0.5)  # Scale action\n                    return base_price * price_multiplier\n                except Exception as e:\n                    logging.error(f\"Q-learning prediction error: {e}\")\n                    # Fallback to simple rule-based pricing\n                    return 10.0 * (1.0 + predicted_occupancy)\n            \n            # Simple rule-based fallback\n            return 10.0 * (1.0 + row['occupancy_rate'])\n        \n        # Apply predictions\n        predictions_table = features_table.with_columns(\n            predicted_occupancy=pw.apply(predict_occupancy, features_table),\n            temporal_pattern=pw.apply(predict_temporal_pattern, features_table)\n        )\n        \n        # Apply price optimization\n        predictions_table = predictions_table.with_columns(\n            optimal_price=pw.apply(optimize_price, predictions_table)\n        )\n        \n        return predictions_table\n    \n    def setup_output_stream(self, predictions_table: pw.Table, output_sink: str):\n        \"\"\"Setup output stream for pricing decisions\"\"\"\n        if output_sink.startswith('kafka://'):\n            pw.io.kafka.write(\n                predictions_table.select(\n                    pw.this.location_id,\n                    pw.this.timestamp, \n                    pw.this.predicted_occupancy,\n                    pw.this.optimal_price\n                ),\n                rdkafka_settings={\n                    'bootstrap.servers': output_sink.replace('kafka://', '')\n                },\n                topic=self.config['output_topic'],\n                format='json'\n            )\n        elif output_sink.endswith('.csv'):\n            pw.io.csv.write(\n                predictions_table.select(\n                    pw.this.location_id,\n                    pw.this.timestamp,\n                    pw.this.predicted_occupancy, \n                    pw.this.optimal_price\n                ),\n                output_sink\n            )\n    \n    def register_models(self, models: Dict[str, Any]):\n        \"\"\"Register ML models for inference\"\"\"\n        self.models = models\n        logging.info(f\"Registered {len(models)} models: {list(models.keys())}\")\n    \n    def run_streaming_pipeline(self, input_source: str, output_sink: str):\n        \"\"\"Run the complete streaming pipeline\"\"\"\n        logging.info(\"Starting Pathway streaming pipeline...\")\n        \n        # Setup input stream\n        input_table = self.setup_input_stream(input_source)\n        \n        # Feature engineering\n        features_table = self.feature_engineering_pipeline(input_table)\n        \n        # ML inference\n        predictions_table = self.ml_inference_pipeline(features_table)\n        \n        # Setup output\n        self.setup_output_stream(predictions_table, output_sink)\n        \n        # Run the pipeline\n        pw.run()\n\n\nclass ParkingDataSchema(pw.Schema):\n    \"\"\"Schema for parking data stream\"\"\"\n    location_id: str\n    timestamp: pw.DateTimeNaive\n    occupancy: int\n    capacity: int\n    queue_length: int\n    traffic_numeric: int\n    vehicle_numeric: int\n    is_special_day: bool\n\n\n# Usage example:\n# streaming_pipeline = ParkingDataStream()\n# streaming_pipeline.register_models({\n#     'xgboost_forecaster': xgboost_model,\n#     'lstm_model': lstm_model, \n#     'q_agent': q_learning_agent\n# })\n# streaming_pipeline.run_streaming_pipeline(\n#     input_source='kafka://localhost:9092',\n#     output_sink='kafka://localhost:9092'\n# )
