requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# Traditional Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning for LSTM/GRU
tensorflow>=2.8.0
torch>=1.10.0
torchvision>=0.11.0
keras>=2.8.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.6.0
tianshou>=0.4.8

# Time Series Analysis
sktime>=0.13.0
tslearn>=0.5.2
statsforecast>=1.0.0

# Streaming & Real-time
pathway-python>=0.1.0

# Visualization
bokeh>=2.4.0
plotly>=5.0.0
dash>=2.0.0

# Data Processing
scipy>=1.7.0
statsmodels>=0.13.0

# Development Tools
pytest>=6.0.0
black>=21.0.0
flake8>=4.0.0
pre-commit>=2.15.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
click>=8.0.0
